requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
PyMySQL
Pytest
